chore(bemHelper): move to utils, not a component
