--- components/BemHelper.js
@@ -1,13 +1,0 @@
-function BemHelper(block) {
-  return function(element, modifier) {
-    if (!element) {
-      return block;
-    }
-    if (!modifier) {
-      return block + '--' + element;
-    }
-    return block + '--' + element + '__' + modifier;
-  };
-}
-
-module.exports = BemHelper;
