+++ lib/utils.js
@@ -61,7 +61,20 @@ function renderTemplate({template, templateHelpers, data}) {
   return content;
 }
 
+function bemHelper(block) {
+  return function(element, modifier) {
+    if (!element) {
+      return block;
+    }
+    if (!modifier) {
+      return block + '--' + element;
+    }
+    return block + '--' + element + '__' + modifier;
+  };
+}
+
 module.exports = {
-  getContainerNode: getContainerNode,
-  renderTemplate: renderTemplate
+  getContainerNode,
+  renderTemplate,
+  bemHelper
 };
